twrp.fstab : Update locations
Pico mostly uses /sd-ext as /data owing to lack of enough NAND. Automatic TWRP-compatible backup apps backup the /data mountpoint which is essentially a stub and doesn't contain any actual data. This commit simply swaps the locations to /data and /sd-ext so that the stub NAND partition is now /sd-ext and /dev/block/mmcblk0p2 is /data, like it is in the actual ROMs

Patch-Set 2: Change location of /cache as well. This device uses PLL2 OC, which burns off any NAND that's mounted R/W . Moving the /cache to an external sd partition prevents irrecoverable damage to the NAND. Most ROMs use this fstab layout.
Change-Id: I529cacf1ec2b274499e12794e9fb476631401ae0
Signed-off-by: MSF Jarvis <msfjarvis@gmail.com>